Started using monad transformers in Lambda.
I'm not sure that this is a good idea. The definitional interpreter is
perhaps a little nicer (except for the awkward treatment of later),
and switching to a different monad is perhaps easier, but on the other
hand the proofs are perhaps a little more complicated than before, and
I added quite a bit of new infrastructure.