Removed a special case from the virtual machine.
The case that was removed was the one where a tail call was made when
there was only a single value on the stack. I was not entirely happy
with this special case, and Ulf Norell reacted negatively to it when
he saw a presentation of this code.

I also tweaked the "done" case.